+++ components/Chart/ChartList.tsx
@@ -1,25 +1,11 @@
-import testPhoto from '../../testEKGImage.jpeg'
 import { LayerType } from './Chart'
+import ChartObject from './ChartObject'
 
 interface ChartListProps {
   layers: Array<LayerType>
 }
 
-interface ChartObjProps {
-  layerId: string
-}
-
 const ChartList = (props: ChartListProps) => {
-  // props.layers.forEach((layer: LayerType) => {
-  //   if (layer.selected) {
-  //     chartComponentList.push(
-  //       <div key={layer.id} className="ChartContainer">
-  //         <ChartObj layerId={layer.id} />
-  //       </div>
-  //     )
-  //   }
-  // })
-
   // TODO make this interact with "toggle chart components" so we display the right components
   // of the EKG
   const dummyChartComponentData = [
@@ -33,7 +19,7 @@ const ChartList = (props: ChartListProps) => {
   while (i < 3) {
     chartComponentList.push(
       <div key={dummyChartComponentData[i].name} className="ChartContainer">
-        <ChartObj layerId={dummyChartComponentData[i].name} />
+        <ChartObject layerId={dummyChartComponentData[i].name} />
       </div>
     )
     i += 1
@@ -42,20 +28,6 @@ const ChartList = (props: ChartListProps) => {
   return (
     <div style={{ display: 'flex', flexDirection: 'column' }}>
       {chartComponentList}
-    </div>
-  )
-}
-
-const ChartObj = ({ layerId }: ChartObjProps) => {
-  // const [state, dispatch] = useReducer(oldState, action) // THIS IS FOR PANNING/ZOOMING
-  // zander is so sexy 🤤
-
-  return (
-    <div className="ChartObj">
-      <p className="Subheading" style={{ margin: '10px' }}>
-        {layerId}
-      </p>
-      <img src={testPhoto.src} />
     </div>
   )
 }
